==> Lab3_Computational_Text_Analysis_I/part3_test_hypothesis.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import re
import nltk

DATA_DIR = Path("../data")
data_path = DATA_DIR / "tweets_account_Advocacy_Political.xlsx"
df = pd.read_excel(data_path)
df.head()

df["Type"].value_counts()

df.columns

## 1. Clean  Data

# 清洗推文文本的函数
def clean_tweet(text):
    if pd.isna(text):
        return ""

    # 1. 移除 RT @xxx:
    text = re.sub(r"^RT\s+@\w+[\:\.\,\!\?\-\;]*\s*", "", text)
    # 2. 移除 @用户名
    text = re.sub(r"[\.\(\)\[\]\{\}!?,;:'\"\-]*@\w+[\.\(\)\[\]\{\}!?,;:'\"\-]*", "", text)


    # 3. 去掉 URL
    text = re.sub(r"http\S+|www\.\S+", "", text)

    # # 4. 去掉 # 号但保留 hashtag 单词
    # text = text.replace("#", "")

    # === 处理 hashtag ===
    # 1）保留“在句子中的 hashtag” → 只去 #
    text = re.sub(r"(?<![A-Za-z])#([A-Za-z0-9_]+)[\.\,\!\?\:\;\-]*", "", text)

    # 2）删除“独立标签 hashtag”
    text = re.sub(r"(?<![A-Za-z])#([A-Za-z0-9_]+)", "", text)

    # 5. 多个空格合并为一个空格
    text = re.sub(r"\s+", " ", text).strip()

    return text


# 查看是否有重复文本数据
def duplicate_texts(df,col):
    len(df['tweet_text'].unique())
    print(f"总推文数: {len(df)}, 唯一推文数: {len(df[col].unique())}") 
    
    df = df.copy()
    df= df.sort_values(by=['author_id', 'posting_date'])
    return df.drop_duplicates(subset=['author_id', col], keep='last')


print(f"原始推文数: {df.shape}")
# ① 过滤英文
df_eng = df.loc[df["language"] == "en"]
print(f"英文推文数: {df_eng.shape}")

# ② 删除重复文本
df_dups = duplicate_texts(df_eng, "tweet_text")
print(f"去重后英文推文数: {df_dups.shape}")

# ③ 清洗 tweet_text
df_dups["clean_text"] = df_dups["tweet_text"].map(clean_tweet)
print(f"清洗后英文推文数: {df_dups.shape}")


## 2. Test sentiment variance 
* Advocacy actors show higher sentiment variance (stronger emotional extremes), while political actors produce more sentimentally stable content.
* 倡导类账号的情绪更极端，而政治类账号情绪更稳定

from nltk.sentiment.vader import SentimentIntensityAnalyzer
nltk.download('vader_lexicon')

sid = SentimentIntensityAnalyzer()

def predict_vader(text):
    score = sid.polarity_scores(text)['compound']
    if score > 0.05:
        return 1
    elif score < -0.05:
        return -1
    return 0

df_dups['pred_vader'] = df_dups["clean_text"].apply(predict_vader)

# 按 Type 统计情绪标签数量
df_dups.groupby("Type")['pred_vader'].value_counts()


# 计算两组的方差
df_dups.groupby("Type")['pred_vader'].var()


import seaborn as sns
import matplotlib.pyplot as plt

sns.countplot(data=df_dups, x='pred_vader', hue='Type')
plt.show()
